Added HOC to add click tracker
